core/requirements.txt
chromadb==1.0.15
fastapi==0.116.1
httpx==0.28.1
numpy==2.3.2
openai==1.97.1
protobuf==6.31.1
pydantic==2.11.7
pydantic_settings==2.10.1
pymongo==4.13.2
python-decouple==3.8
python-dotenv==1.1.1
qdrant_client==1.15.0
sentence_transformers==5.0.0
supabase==2.17.0
uvicorn==0.35.0
